Rename HTMLParser to html_parser
